requirements.txt
streamlit==1.30.0
pypdf==4.0.0
google-generativeai==0.3.2
openai==1.10.0
python-dotenv==1.0.0
